34 all buttons disable when performing actions to avoid double posting (#35)
* Updated buttons with network calls to have loading indicator

* Handle auth persistant correctly

+++ web/src/components/TheDashboard.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import type { GetRecentStatistics200Response } from '@/repositories/clients/private';
-import { auth } from '@/services/firebase';
 import { MyPrivateApi } from '@/services/requests-private';
 import { useGameStore } from '@/stores/game-store';
 import { useUserStore } from '@/stores/user-store';
-import { signOut } from 'firebase/auth';
+import { signOut, type Auth } from 'firebase/auth';
 import { DateTime } from 'luxon';
-import { computed, onMounted, reactive, ref } from 'vue';
+import { computed, inject, onMounted, reactive, ref } from 'vue';
+import ButtonComponent from './ButtonComponent.vue';
 
 const errorMessage = ref('')
 const userEmail = ref('')
@@ -22,7 +22,8 @@ const avgTimePerPointWonDecorated = ref("")
 const avgTimePerPointLostDecorated = ref("")
 const avgTimePerGameDecorated = ref("")
 const showRecentStatisticsHint = ref(false)
-
+const isLoading = ref(false)
+const auth = inject<Auth>("auth");
 
 const userStore = useUserStore()
 userStore.setBackendUrl(import.meta.env.VITE_PROXY_URL)
@@ -36,16 +37,22 @@ onMounted(async () => {
 })
 
 const submitLogout = async () => {
-    try {
-        await signOut(auth)
-        const privApi = new MyPrivateApi(import.meta.env.VITE_PROXY_URL)
-        privApi.deleteSession()
+    isLoading.value = true
 
+    try {
+        if (auth) {
+            await signOut(auth)
+            const privApi = new MyPrivateApi(import.meta.env.VITE_PROXY_URL)
+            privApi.deleteSession()
+            isLoading.value = false
+        } else {
+            errorMessage.value = "no auth to logout"
+            isLoading.value = false
+        }
     } catch (error: any) {
         errorMessage.value = error.message
-
+        isLoading.value = false
     }
-    errorMessage.value = ""
 }
 
 const getUserEmail = async () => {
@@ -153,10 +160,11 @@ const decorateSecondsWithHour = (seconds: number) => {
 
             </div>
 
-            <form @submit.prevent="submitLogout">
-
+            <form>
                 <div class="actions">
-                    <button type="submit" class="button button-secondary">Logout</button>
+                    <ButtonComponent type="secondary" :isLoading="isLoading" @click.prevent="submitLogout">
+                        Logout
+                    </ButtonComponent>
                 </div>
 
                 <div v-if="errorMessage" class="error">
@@ -172,9 +180,9 @@ const decorateSecondsWithHour = (seconds: number) => {
                 </button>
             </RouterLink> -->
             <RouterLink to="/game/setup" class="mt">
-                <button class="button button-primary">
+                <ButtonComponent type="primary">
                     New Game
-                </button>
+                </ButtonComponent>
             </RouterLink>
             <div class="recent-statistics" v-if="showRecentStatisticsHint">
                 <h5>Recent Statistics</h5>
